Added deleteAssignment method in AssignmentController

+++ cst438_gradebook/src/main/java/com/cst438/controllers/AssignmentController.java
@@ -1,6 +1,7 @@
 package com.cst438.controllers;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,11 +13,14 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.cst438.domain.Assignment;
 import com.cst438.domain.AssignmentDTO;
+import com.cst438.domain.AssignmentGrade;
+import com.cst438.domain.AssignmentGradeRepository;
 import com.cst438.domain.AssignmentRepository;
 import com.cst438.domain.Course;
 import com.cst438.domain.CourseRepository;
@@ -30,6 +34,9 @@ public class AssignmentController {
 	
 	@Autowired
 	CourseRepository courseRepository;
+	
+	@Autowired
+	AssignmentGradeRepository assignmentGradeRepository;
 	
 	@GetMapping("/assignment")
 	public AssignmentDTO[] getAllAssignmentsForInstructor() {
@@ -50,9 +57,10 @@ public class AssignmentController {
 		return result;
 	}
 	
-	//create assignment (POST requests)
+	
+	//create assignment (POST)
 	@PostMapping("/assignment")
-	public Assignment createAssignment(@RequestBody AssignmentDTO assignmentDTO) {
+	public int createAssignment(@RequestBody AssignmentDTO assignmentDTO) {
 		//first convert the AssignmentDTO to Assignment entity
 		Assignment assignment = new Assignment();
 		assignment.setName(assignmentDTO.assignmentName());
@@ -65,53 +73,94 @@ public class AssignmentController {
 		}
 		assignment.setCourse(course);
 		
-		//save and return the created assignment
-		return assignmentRepository.save(assignment);
+		//save and return the ID of the created assignment
+		assignment = assignmentRepository.save(assignment);
+		return assignment.getId();
 	}
 	
-	//Retrieve a single assignment (GET)
+	
+	//Retrieve assignment by primary key (GET)
 	@GetMapping("/assignment/{id}")
-	public Assignment getAssignmentById(@PathVariable int id) {
+	public AssignmentDTO getAssignment(@PathVariable("id") int id) {
 		Assignment assignment = assignmentRepository.findById(id).orElse(null);
-        if (assignment == null) {
+        
+		//exception if no assignment with given ID is found.
+		if (assignment == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Assignment ID not found");
         }
+		
+		//format the date in YYYY MM DD
+		String formattedDate = assignment.getDueDate().toLocalDate().toString();
 
-        return assignment;
+		//create an assignmentDTO object and set its fields based on the Assignment object
+        AssignmentDTO assignmentDTO = new AssignmentDTO(
+        		assignment.getId(),
+        		assignment.getName(),
+        		formattedDate,
+        		assignment.getCourse().getTitle(),
+        		assignment.getCourse().getCourse_id()
+        		);
+
+		//return AssignmentDTO object
+		return assignmentDTO;
     }
 	
 	//Update (PUT)
 	@PutMapping("/assignment/{id}")
-	public Assignment updateAssignment(@PathVariable int id, @RequestBody AssignmentDTO assignmentDTO) {
-		//check if assignment exists
+	public void updateAssignment(@PathVariable("id") int id, @RequestBody AssignmentDTO assignmentDTO) {
+		//check if assignment exists using the repository
 		Assignment assignment = assignmentRepository.findById(id).orElse(null);
-        if (assignment == null) {
+        
+		//throw exception if assignment doesn't exists
+		if (assignment == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Assignment not found");
         }
 
         assignment.setName(assignmentDTO.assignmentName());
         assignment.setDueDate(java.sql.Date.valueOf(assignmentDTO.dueDate()));
         
+        //check if course exists
         Course course = courseRepository.findById(assignmentDTO.courseId()).orElse(null);
+        
         if (course == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Course not found, try another one");
         }
+        
+        //set the course for the assignment
         assignment.setCourse(course);
-
-        return assignmentRepository.save(assignment);
+        
+        //save the updated assignment
+        assignmentRepository.save(assignment);
     }
 	
-	//Delete (DELETE)
+	//delete (DELETE)
+	//delete /assignment/{id}
+	//delete /assignment/{id}?force=yes
 	@DeleteMapping("/assignment/{id}")
-	public void deleteAssignment(@PathVariable int id) {
-		//check if assignment exists or return an error message
-		 Assignment assignment = assignmentRepository.findById(id).orElse(null);
-	        if (assignment == null) {
-	            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Assignment not found");
-	        }
-
-	        assignmentRepository.delete(assignment);
+	public void deleteAssignment(@PathVariable int id, @RequestParam("force") Optional<String> force) {
+		//check if assignment exists with repository or return an error message
+		Assignment assignment = assignmentRepository.findById(id).orElse(null);
+	     
+		if (assignment == null) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Assignment not found");
 	    }
-	
-	//TODO get assignment grades method
+		//get assignment grades
+		Iterable<AssignmentGrade> grades = assignmentGradeRepository.findAll();
+		//check
+		boolean hasGrades=false;
+		for (AssignmentGrade grade : grades) {
+			if (grade.getAssignment().getId() == id) {
+				hasGrades=true;
+				break;
+			}
+		}
+		//if assignment has grades and no parameter is used
+		if(hasGrades && force.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.CONFLICT, "Assignment has grades!, if you want to continue use force parameter");
+		}
+		//delete the assignment
+	    assignmentRepository.delete(assignment);
+	    
+	    
+		}
 }
